:recycle: Split account processor into directory module by command and query
Pure mechanical move of adapter/src/processor/account.rs into a directory module; no behavior change.
