+++ adapter/src/processor/account/mod.rs
@@ -1,0 +1,8 @@
+mod command;
+mod query;
+
+pub use command::{
+    AccountCommandProcessor, CreateAccountParam, DependOnAccountCommandProcessor,
+    DependOnAccountSignal, UpdateAccountParam,
+};
+pub use query::{AccountQueryProcessor, DependOnAccountQueryProcessor};
